# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "HOTEL",
        "description": "API for hotel management",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/payments": {
            "get": {
                "operationId": "allPayments",
                "tags": [
                    "Payments"
                ],
                "description": "Recupere la liste de tous les paiements",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_payment": {
                                        "type": "integer"
                                    },
                                    "payment_date": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "payment_method": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "404": {
                    "description": "Not found"
                },
                "500": {
                    "description": "Internal server error"
                }
            },
            "post": {
                "operationId": "addPayment",
                "tags": [
                    "Payments"
                ],
                "description": "Ajoute un paiement",
                "parameters": [
                    {
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "payment_date": {
                                    "type": "string"
                                },
                                "amount": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "payment_method": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/payments/{id}": {
            "get": {
                "operationId": "payment_id",
                "tags": [
                    "Payments"
                ],
                "description": "Recupere un paiement par son id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id_payment": {
                                    "type": "integer"
                                },
                                "payment_date": {
                                    "type": "string"
                                },
                                "amount": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "payment_method": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "patch": {
                "operationId": "updatePayment",
                "tags": [
                    "Payments"
                ],
                "description": "Met a jour un paiement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "payment_date": {
                                    "type": "string"
                                },
                                "amount": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "payment_method": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "operationId": "deletePayment",
                "tags": [
                    "Payments"
                ],
                "description": "Supprime un paiement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/payments/countByMethod/{method}": {
            "get": {
                "operationId": "countByMethod",
                "tags": [
                    "Payments"
                ],
                "description": "Recupere le nombre de reservations par methode de paiement",
                "parameters": [
                    {
                        "name": "method",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id_payment": {
                                    "type": "integer"
                                },
                                "payment_date": {
                                    "type": "string"
                                },
                                "amount": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "payment_method": {
                                    "type": "string"
                                },
                                "id_reservation": {
                                    "type": "integer"
                                },
                                "id_client": {
                                    "type": "integer"
                                },
                                "id_room": {
                                    "type": "integer"
                                },
                                "checkin_date": {
                                    "type": "string"
                                },
                                "checkout_date": {
                                    "type": "string"
                                },
                                "total_price": {
                                    "type": "number",
                                    "format": "float"
                                },
                                "reservation_status": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/inclure/countServiceByServiceName/{service_name}": {
            "get": {
                "operationId": "countServiceByServiceName",
                "tags": [
                    "Inclure"
                ],
                "description": "Recupere le nombre de services par nom de service",
                "parameters": [
                    {
                        "name": "service_name",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "total": {
                                    "type": "number",
                                    "format": "float"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    }
}